Move structures used in lstm_d to a separate file.
